backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pdfplumber
python-docx
openpyxl
pytesseract
pillow
pandas
azure-storage-blob
azure-cosmos
psycopg2-binary
azure-ai-documentintelligence
reportlab
openai
httpx
aiofiles
azure-identity
PyJWT
passlib[bcrypt]
python-dotenv
email-validator
